requirements.txt
jsonschema
flask
flask-cors
PyYAML
waitress
werkzeug==3.0.1
psutil